requirements.txt
qdrant_client
psycopg2-binary
sentence-transformers
python-dotenv